progress on lgl and lgr figures
